Rename elbowarrow.ts
